retrieval/requirements.txt
fastapi
uvicorn
azure-core
azure-search-documents
openai
python-dotenv
